HKP: Fixed handling of empty proxy string.
Previously, CURLOPT_PROXY was set to <connection_number>:password@PROXY.
However, when PROXY is an empty string this does not result in CURL's proxy
functionality being disabled, and so the request hangs.

We have added an check for this, and now explicitly set CURLOPT_PROXY to ""
when an empty SOCKS5 host is supplied.  While this will not normally be
the case, as the software is meant to be used with Tor, operation without a
proxy can be useful for testing purposes.