daemon: Removed guard interval (fixes #2).
Previously, between each request interval there was a 20% guard interval.  This
was necessary because we always placed the next request into the next window,
and if the last request spilled over into the following window then it would
be skipped.

But a guard interval reduces anonymity if the clocks are not well-synchronised,
so it must be removed.  This requires a refactor of the timing code,
though given its previous horror this was inevitable anyway.  Now the
request windows always follow one another, even if this means that a spillover
will sometimes result in us trying to schedule a request in the past.