Build: Added Clang compiler flags.
We were blindly assuming that gcc was in use, which is obviously poor form.
CMakeLists.txt now checks for GCC and Clang, setting their flags if and only if
they are in use.

It seems that boost::spirit::qi needed a greater maximum template depth than
Clang allows by default (500), so we have added -ftemplate-depth=1024.